requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
psycopg[binary]>=3.2,<4.0
gunicorn>=23.0,<24.0
python-dotenv>=1.0,<2.0
